% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_issue_comment.R
\name{on_issue_comment}
\alias{on_issue_comment}
\title{Generate a GitHub Actions 'on:' specification for the 'issue_comment' event}
\usage{
on_issue_comment(types = c("created", "edited", "deleted"))
}
\arguments{
\item{types}{Character vector of issue_comment event types to trigger. See Details for choices. Default is c("created", "edited", "deleted").}
}
\value{
A named list representing the YAML 'on:' block for the issue_comment event.
}
\description{
Returns a list ready to be converted to a YAML block for the 'on:' section of a GitHub Actions workflow.
}
\details{
Valid types for issue_comment are: "created", "edited", "deleted". The default is all three.
}
\examples{
on_issue_comment()
on_issue_comment(types = c("created"))
}
