% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_issues.R
\name{on_issues}
\alias{on_issues}
\title{Generate a GitHub Actions 'on:' specification for the 'issues' event}
\usage{
on_issues(
  types = c("opened", "edited", "deleted", "transferred", "pinned", "unpinned", "closed",
    "reopened", "assigned", "unassigned", "labeled", "unlabeled", "locked", "unlocked",
    "milestoned", "demilestoned")
)
}
\arguments{
\item{types}{Character vector of issue event types to trigger.}
}
\value{
A named list representing the YAML 'on:' block for the issues event.
}
\description{
Returns a list ready to be converted to a YAML block to use in the 'on:' block in a GitHub Actions workflow
}
\examples{
on_issues()
on_issues(types = c('opened', 'edited'))
}
