% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_pull_request_review.R
\name{on_pull_request_review}
\alias{on_pull_request_review}
\title{Generate a GitHub Actions 'on:' specification for the 'pull_request_review' event}
\usage{
on_pull_request_review(types = c("submitted", "edited", "dismissed"))
}
\arguments{
\item{types}{Character vector of pull_request_review event types to trigger. See Details for choices. Default is c("submitted", "edited", "dismissed").}
}
\value{
A named list representing the YAML 'on:' block for the pull_request_review event.
}
\description{
Returns a list ready to be converted to a YAML block for the 'on:' section of a GitHub Actions workflow.
}
\details{
Valid types for pull_request_review are: "submitted", "edited", "dismissed". The default is all three.
}
\examples{
on_pull_request_review()
on_pull_request_review(types = c("submitted"))
}
