% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_pull_request_review_comment.R
\name{on_pull_request_review_comment}
\alias{on_pull_request_review_comment}
\title{Generate a GitHub Actions 'on:' specification for the 'pull_request_review_comment' event}
\usage{
on_pull_request_review_comment(types = c("created", "edited", "deleted"))
}
\arguments{
\item{types}{Character vector of pull_request_review_comment event types to trigger. See Details for choices. Default is c("created", "edited", "deleted").}
}
\value{
A named list representing the YAML 'on:' block for the pull_request_review_comment event.
}
\description{
Returns a list ready to be converted to a YAML block for the 'on:' section of a GitHub Actions workflow.
}
\details{
Valid types for pull_request_review_comment are: "created", "edited", "deleted". The default is all three.
}
\examples{
on_pull_request_review_comment()
on_pull_request_review_comment(types = c("created"))
}
