% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_release.R
\name{on_release}
\alias{on_release}
\title{Generate a GitHub Actions 'on:' specification for the 'release' event}
\usage{
on_release(
  types = c("published", "unpublished", "created", "edited", "deleted", "prereleased",
    "released")
)
}
\arguments{
\item{types}{Character vector of release event types to trigger. See Details for choices. Default is c("published", "unpublished", "created", "edited", "deleted", "prereleased", "released").}
}
\value{
A named list representing the YAML 'on:' block for the release event.
}
\description{
Returns a list ready to be converted to a YAML block for the 'on:' section of a GitHub Actions workflow.
}
\details{
Valid types for release are: "published", "unpublished", "created", "edited", "deleted", "prereleased", "released". The default is all seven.
}
\examples{
on_release()
on_release(types = c("published", "created"))
}
