% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_browse_url.R
\name{tool_browse_url}
\alias{tool_browse_url}
\title{Tool: Browse URL}
\usage{
tool_browse_url()
}
\description{
Returns a tool object for getting the current content of the system clipboard as a string.
}
