% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_clipboard_content.R
\name{tool_clipboard_content}
\alias{tool_clipboard_content}
\title{Tool: Clipboard content}
\usage{
tool_clipboard_content()
}
\description{
Returns a tool object for getting the current content of the system clipboard as a string.
}
