% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_cursor_position.R
\name{tool_cursor_position}
\alias{tool_cursor_position}
\title{Tool: Mouse cursor position}
\usage{
tool_cursor_position()
}
\description{
Returns a tool object for getting the current mouse cursor position.
}
