% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_file_copy.R
\name{tool_file_copy}
\alias{tool_file_copy}
\title{Tool: File copy}
\usage{
tool_file_copy(ask = TRUE)
}
\arguments{
\item{ask}{Boolean. If \code{TRUE} (default), ask for confirmation before copying the file.}
}
\description{
Returns a tool object for copying a file from one path to another.
}
