% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_file_delete.R
\name{tool_file_delete}
\alias{tool_file_delete}
\title{Tool: Delete a file}
\usage{
tool_file_delete(ask = TRUE)
}
\arguments{
\item{ask}{Boolean. If \code{TRUE} (default), ask for confirmation before deleting the file.}
}
\description{
Returns a tool object for deleting a specified file.
}
\examples{
\dontrun{
  chat <- ellmer::chat_openai()
  chat$register_tool(tool_file_delete())
  tmp <- tempfile()
  writeLines("delete me", tmp)
  chat$chat(paste("Delete the file", tmp))
}
}
