% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_ga_api_delete.R
\name{tool_ga_api_delete}
\alias{tool_ga_api_delete}
\title{Tool: GitHub API DELETE}
\usage{
tool_ga_api_delete()
}
\description{
Makes an authenticated DELETE call to the GitHub REST API using the GITHUB_TOKEN provided by the runner.
}
