% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_ga_api_patch.R
\name{tool_ga_api_patch}
\alias{tool_ga_api_patch}
\title{Tool: GitHub API PATCH}
\usage{
tool_ga_api_patch()
}
\description{
Makes an authenticated PATCH call to the GitHub REST API using the GITHUB_TOKEN provided by the runner.
}
