% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_ga_api_post.R
\name{tool_ga_api_post}
\alias{tool_ga_api_post}
\title{Tool: GitHub API POST}
\usage{
tool_ga_api_post()
}
\description{
Makes an authenticated POST call to the GitHub REST API using the GITHUB_TOKEN provided by the runner.
}
