% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_ga_api_put.R
\name{tool_ga_api_put}
\alias{tool_ga_api_put}
\title{Tool: GitHub API PUT}
\usage{
tool_ga_api_put()
}
\description{
Makes an authenticated PUT call to the GitHub REST API using the GITHUB_TOKEN provided by the runner.
}
