% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_ga_event_context.R
\name{tool_ga_event_context}
\alias{tool_ga_event_context}
\title{Tool: GA Event Context}
\usage{
tool_ga_event_context()
}
\value{
A tool object for use in agentic agents.
}
\description{
Returns a tool object for retrieving a value from the github actions event payload.
}
\examples{
chat <- ellmer::chat_openai()
chat$register_tool(tool_ga_event_context())
chat$chat("What's the current issue's body?")
}
