% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_image_metadata.R
\name{tool_image_metadata}
\alias{tool_image_metadata}
\title{Tool: Image metadata}
\usage{
tool_image_metadata()
}
\description{
Returns a tool object for getting image metadata (width, height, format, colorspace, filesize).
}
