% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_keypress.R
\name{tool_keypress}
\alias{tool_keypress}
\title{Tool: Keypress}
\usage{
tool_keypress(ask = TRUE)
}
\arguments{
\item{ask}{Boolean. If \code{TRUE} (default), ask for confirmation before simulating the keypress.}
}
\description{
Returns a tool object for simulating a keypress (with optional modifiers).
}
\examples{
\dontrun{
  chat <- ellmer::chat_openai()
  chat$register_tool(tool_keypress())
  chat$chat("Press the Enter key")
}
}
