% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_left_click.R
\name{tool_left_click}
\alias{tool_left_click}
\title{Tool: Left mouse click}
\usage{
tool_left_click()
}
\description{
Returns a tool object for performing a left mouse click at the given (x, y) screen coordinates.
}
\examples{
\dontrun{
  chat <- ellmer::chat_openai()
  chat$register_tool(tool_left_click())
  chat$chat("Click at position 100, 200")
}
}
