% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_list_files.R
\name{tool_list_files}
\alias{tool_list_files}
\title{Tool: List files and directories}
\usage{
tool_list_files()
}
\description{
Returns a tool object for listing files and directories in a given path.
}
