% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_popup.R
\name{tool_popup}
\alias{tool_popup}
\title{Tool: Popup}
\usage{
tool_popup()
}
\description{
Shows a popup message to the user.
}
\examples{
\dontrun{
  chat <- ellmer::chat_openai()
  chat$register_tool(tool_popup())
  chat$chat("Show a popup saying 'Hello, world!'")
}
}
