% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_right_click.R
\name{tool_right_click}
\alias{tool_right_click}
\title{Tool: Right mouse click}
\usage{
tool_right_click()
}
\description{
Returns a tool object for performing a right mouse click at the given (x, y) screen coordinates.
}
