% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_screen_resolution.R
\name{tool_screen_resolution}
\alias{tool_screen_resolution}
\title{Tool: Screen resolution}
\usage{
tool_screen_resolution()
}
\description{
Returns a tool object for getting the screen resolution (width and height in pixels).
}
