% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_session_info.R
\name{tool_session_info}
\alias{tool_session_info}
\title{Tool: R session info}
\usage{
tool_session_info()
}
\description{
Returns a tool object for getting the R session information as a string.
}
\examples{
\dontrun{
  chat <- ellmer::chat_openai()
  chat$register_tool(tool_session_info())
  chat$chat("Show the R session info")
}
}
