% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_system_info.R
\name{tool_system_info}
\alias{tool_system_info}
\title{Tool: System info}
\usage{
tool_system_info()
}
\description{
Returns a tool object for getting the system information as a string.
}
\examples{
\dontrun{
  chat <- ellmer::chat_openai()
  chat$register_tool(tool_system_info())
  chat$chat("Show the system info")
}
}
