% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_working_directory.R
\name{tool_working_directory}
\alias{tool_working_directory}
\title{Tool: Working directory}
\usage{
tool_working_directory()
}
\description{
Returns a tool object for getting the current working directory.
}
