% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_write_to_file.R
\name{tool_write_to_file}
\alias{tool_write_to_file}
\title{Tool: Write to File}
\usage{
tool_write_to_file()
}
\value{
A tool object
}
\description{
Returns a tool for writing a character string to a flat file at the specified path.
}
\examples{
tool <- tool_write_to_file()
tool$run(path = tempfile(), content = "hello")
}
