% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{agent}
\alias{agent}
\title{Create an agent with all agentic tools}
\usage{
agent(system_prompt = NULL, ..., model = NULL)
}
\arguments{
\item{system_prompt}{The system prompt.
If not provided and if a file named \code{agentic-rules.md} is found in the working directory
its content will be used as the system prompt.
To force an empty system prompt over 'agentic-rules.md' use \code{system_prompt = ""}.}

\item{...}{Additional arguments passed to the chat function for the provider.}

\item{model}{The model in the form "provider/model", defaults to "openai/gpt-4.1"}
}
\value{
An Chat object with all agentic tools registered.
}
\description{
This function creates an agent using the specified model family (default: 'openai'),
registers all available agentic tools, and returns the agent object.
}
\examples{
ag <- agent()
ag$chat("What time is it?")
}
