% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentic-package.R
\docType{package}
\name{agentic-package}
\alias{agentic}
\alias{agentic-package}
\title{agentic: Tools to make powerful agents}
\description{
A set of functions that can be used to enhance the interaction of LLMs with the system, and in particular tools to use smoothly with the 'ellmer' package.
}
\author{
\strong{Maintainer}: Antoine Fabri \email{antoine.fabri@gmail.com}

}
\keyword{internal}
