% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_copy_to_clipboard.R
\name{tool_copy_to_clipboard}
\alias{tool_copy_to_clipboard}
\title{Tool: Copy to clipboard}
\usage{
tool_copy_to_clipboard()
}
\description{
Returns a tool object for copying text to the system clipboard.
}
\examples{
\dontrun{
  chat <- ellmer::chat_openai()
  chat$register_tool(tool_copy_to_clipboard())
  chat$chat("Copy 'hello world' to the clipboard")
}
}
