% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_double_left_click.R
\name{tool_double_left_click}
\alias{tool_double_left_click}
\title{Tool: Double left mouse click}
\usage{
tool_double_left_click()
}
\description{
Returns a tool object for performing a double left mouse click at the given (x, y) screen coordinates.
}
